add a `minimal` version of PostSummary editor component and remove the `excerpt panel` from inspector controls
