+++ packages/edit-post/src/components/sidebar/post-excerpt/index.js
@@ -1,14 +1,11 @@
 /**
  * WordPress dependencies
  */
-import { __ } from '@wordpress/i18n';
-import { PanelBody } from '@wordpress/components';
 import {
 	PostExcerpt as PostExcerptForm,
 	PostExcerptCheck,
 } from '@wordpress/editor';
-import { compose } from '@wordpress/compose';
-import { withSelect, withDispatch } from '@wordpress/data';
+import { useSelect } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -20,38 +17,18 @@ import { store as editPostStore } from '../../../store';
  */
 const PANEL_NAME = 'post-excerpt';
 
-function PostExcerpt( { isEnabled, isOpened, onTogglePanel } ) {
+export default function PostExcerpt( { isMinimal } ) {
+	const isEnabled = useSelect(
+		( select ) =>
+			select( editPostStore ).isEditorPanelEnabled( PANEL_NAME ),
+		[]
+	);
 	if ( ! isEnabled ) {
 		return null;
 	}
-
 	return (
 		<PostExcerptCheck>
-			<PanelBody
-				title={ __( 'Excerpt' ) }
-				opened={ isOpened }
-				onToggle={ onTogglePanel }
-			>
-				<PostExcerptForm />
-			</PanelBody>
+			<PostExcerptForm isMinimal={ isMinimal } />
 		</PostExcerptCheck>
 	);
 }
-
-export default compose( [
-	withSelect( ( select ) => {
-		return {
-			isEnabled: select( editPostStore ).isEditorPanelEnabled(
-				PANEL_NAME
-			),
-			isOpened: select( editPostStore ).isEditorPanelOpened( PANEL_NAME ),
-		};
-	} ),
-	withDispatch( ( dispatch ) => ( {
-		onTogglePanel() {
-			return dispatch( editPostStore ).toggleEditorPanelOpened(
-				PANEL_NAME
-			);
-		},
-	} ) ),
-] )( PostExcerpt );
