+++ packages/edit-post/src/components/sidebar/post-summary/index.js
@@ -4,17 +4,14 @@
 import { __ } from '@wordpress/i18n';
 import { PanelBody, PanelRow } from '@wordpress/components';
 import { useSelect, useDispatch } from '@wordpress/data';
-import {
-	PostExcerptCheck,
-	PostExcerpt2 as PostExcerpt,
-	PostAuthor,
-} from '@wordpress/editor';
+import { PostAuthor } from '@wordpress/editor';
 
 /**
  * Internal dependencies
  */
 import PostTitle from '../post-title';
 import FeaturedImage from '../featured-image';
+import PostExcerpt from '../post-excerpt';
 import { store as editPostStore } from '../../../store';
 
 /**
@@ -23,24 +20,11 @@ import { store as editPostStore } from '../../../store';
 const PANEL_NAME = 'post-summary';
 
 function PostSummary() {
-	// TODO: check about the below logic and the preferences panels...
-
-	const { isRemoved, isOpened } = useSelect( ( select ) => {
-		// We use isEditorPanelRemoved to hide the panel if it was programatically removed. We do
-		// not use isEditorPanelEnabled since this panel should not be disabled through the UI.
-		const { isEditorPanelRemoved, isEditorPanelOpened } = select(
-			editPostStore
-		);
-		return {
-			isRemoved: isEditorPanelRemoved( PANEL_NAME ),
-			isOpened: isEditorPanelOpened( PANEL_NAME ),
-		};
-	}, [] );
+	const isOpened = useSelect(
+		( select ) => select( editPostStore ).isEditorPanelOpened( PANEL_NAME ),
+		[]
+	);
 	const { toggleEditorPanelOpened } = useDispatch( editPostStore );
-
-	if ( isRemoved ) {
-		return null;
-	}
 	return (
 		<PanelBody
 			className="edit-post-post-summary"
@@ -52,12 +36,7 @@ function PostSummary() {
 			<PanelRow>
 				<PostTitle />
 			</PanelRow>
-			{ /* Todo: I need to check/consolidate with panels preferences */ }
-			<PostExcerptCheck>
-				<PanelRow>
-					<PostExcerpt />
-				</PanelRow>
-			</PostExcerptCheck>
+			<PostExcerpt isMinimal />
 			<PanelRow>
 				<PostAuthor labelPosition="side" />
 			</PanelRow>
