--- packages/editor/src/components/post-excerpt2/check.js
@@ -1,10 +1,0 @@
-/**
- * Internal dependencies
- */
-import PostTypeSupportCheck from '../post-type-support-check';
-
-function PostExcerptCheck( props ) {
-	return <PostTypeSupportCheck { ...props } supportKeys="excerpt" />;
-}
-
-export default PostExcerptCheck;
